Update binaryTree.cpp/hpp&main.cpp - Remove Floats
for more extensive decimal replaced floats with doubles and changed another issue with negatives

+++ main.cpp
@@ -10,7 +10,7 @@ int main(){
     cout << "Enter Arithmetic Equation: ";
     cin >> equation;
     BinaryExpressionTree<string> expressionTree(equation);
-    float answer = expressionTree.evaluate();
+    double answer = expressionTree.evaluate();
     cout << "Result: " << answer << endl; 
     return 0;
 }
